+++ pkg/provider/acceptance_source_load_generator_test.go
@@ -2,6 +2,7 @@ package provider
 
 import (
 	"database/sql"
+	"errors"
 	"fmt"
 	"testing"
 
@@ -419,7 +420,7 @@ func testAccCheckAllSourceLoadGeneratorsDestroyed(s *terraform.State) error {
 		_, err := materialize.ScanSource(db, utils.ExtractId(r.Primary.ID))
 		if err == nil {
 			return fmt.Errorf("SourceLoadGenerator %v still exists", utils.ExtractId(r.Primary.ID))
-		} else if err != sql.ErrNoRows {
+		} else if !errors.Is(err, sql.ErrNoRows) {
 			return err
 		}
 	}