+++ pkg/resources/resource_cluster_replica.go
@@ -3,7 +3,7 @@ package resources
 import (
 	"context"
 	"database/sql"
-	"log"
+	"errors"
 
 	"github.com/MaterializeInc/terraform-provider-materialize/pkg/materialize"
 	"github.com/MaterializeInc/terraform-provider-materialize/pkg/utils"
@@ -57,7 +57,7 @@ func clusterReplicaRead(ctx context.Context, d *schema.ResourceData, meta interf
 		return diag.FromErr(err)
 	}
 	s, err := materialize.ScanClusterReplica(metaDb, utils.ExtractId(i))
-	if err == sql.ErrNoRows {
+	if errors.Is(err, sql.ErrNoRows) {
 		d.SetId("")
 		return nil
 	} else if err != nil {
@@ -139,9 +139,7 @@ func clusterReplicaCreate(ctx context.Context, d *schema.ResourceData, meta inte
 		comment := materialize.NewCommentBuilder(metaDb, o)
 
 		if err := comment.Object(v.(string)); err != nil {
-			log.Printf("[DEBUG] resource failed comment, dropping object: %s", o.Name)
-			b.Drop()
-			return diag.FromErr(err)
+			return rollbackCreate(b, o, "comment", err)
 		}
 	}