+++ pkg/resources/resource_source_table_kafka.go
@@ -3,6 +3,7 @@ package resources
 import (
 	"context"
 	"database/sql"
+	"errors"
 
 	"github.com/MaterializeInc/terraform-provider-materialize/pkg/materialize"
 	"github.com/MaterializeInc/terraform-provider-materialize/pkg/utils"
@@ -317,7 +318,7 @@ func sourceTableKafkaRead(ctx context.Context, d *schema.ResourceData, meta inte
 	}
 
 	t, err := materialize.ScanSourceTableKafka(metaDb, utils.ExtractId(i))
-	if err == sql.ErrNoRows {
+	if errors.Is(err, sql.ErrNoRows) {
 		d.SetId("")
 		return nil
 	} else if err != nil {